feat(refactor) : 3.2 닉네임을 입력하면 전날까지의 크루 출결 횟수를 확인할 수 있다.
- 테스트픽스처를 이용해서 출결 횟수를 바탕으로 테스트 데이터를 만드는 기능 추가
- AttendanceDateTimeFixture 추가
- AttendanceDateTimesFixture 추가

+++ src/main/java/domain/AttendanceHistories.java
@@ -33,13 +33,13 @@ public class AttendanceHistories {
         return attendanceHistories.get(crew);
     }
 
-    public long getPresentCount(Crew crew, LocalDate toDate) {
+    public int getPresentCount(Crew crew, LocalDate toDate) {
         validateCrewPresence(crew);
         AttendanceDateTimes attendanceDateTimes = attendanceHistories.get(crew);
         return attendanceDateTimes.getPresentCount(toDate);
     }
 
-    public long getTardyCount(Crew crew, LocalDate toDate) {
+    public int getTardyCount(Crew crew, LocalDate toDate) {
         validateCrewPresence(crew);
         AttendanceDateTimes attendanceDateTimes = attendanceHistories.get(crew);
         return attendanceDateTimes.getTardyCount(toDate);
